@@ -1,7 +1,7 @@
 package com.xingren.excel.util;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 import java.time.LocalDateTime;
 import java.time.OffsetDateTime;
@@ -17,7 +17,7 @@ public class DateUtilTest {
     public void test_stringOffsetDateTimeConvert() {
         OffsetDateTime offsetDateTime = DateUtil.parseToOffsetDateTime("2020/02/10", "yyyy/MM/dd");
         String time = DateUtil.formatOffsetDateTime(offsetDateTime, "yyyy/MM/dd");
-        Assert.assertTrue(time.equals("2020/02/10"));
+        Assertions.assertTrue(time.equals("2020/02/10"));
 
     }
 
@@ -25,7 +25,7 @@ public class DateUtilTest {
     public void test_parseToOffsetDateTime() {
         LocalDateTime localDateTime = DateUtil.parseToLocalDateTime("2020/02/10", "yyyy/MM/dd");
         String time = DateUtil.formatLocalDateTime(localDateTime, "yyyy/MM/dd");
-        Assert.assertEquals("2020/02/10", time);
+        Assertions.assertEquals("2020/02/10", time);
 
     }
 
@@ -34,7 +34,7 @@ public class DateUtilTest {
         Date date = new Date();
         LocalDateTime localDateTime = DateUtil.dateToLocalDateTime(date);
         Date parseDate = DateUtil.localDateTimeToDate(localDateTime);
-        Assert.assertEquals(date.getTime(),parseDate.getTime());
+        Assertions.assertEquals(date.getTime(), parseDate.getTime());
 
     }
 
