@@ -1,13 +1,13 @@
 package com.xingren.excel.util;
 
-import org.junit.Assert;
-import org.junit.Test;
+import org.junit.jupiter.api.Assertions;
+import org.junit.jupiter.api.Test;
 
 public class StrUtilTest {
 
     @Test
     public void test_captureName() {
-        Assert.assertTrue(StrUtil.upperFirstChar("name").equals("Name"));
-        Assert.assertTrue(StrUtil.upperFirstChar("Name").equals("Name"));
+        Assertions.assertTrue(StrUtil.upperFirstChar("name").equals("Name"));
+        Assertions.assertTrue(StrUtil.upperFirstChar("Name").equals("Name"));
     }
 }
