@@ -1,6 +1,7 @@
 package com.xingren.excel.converter.read.impl;
 
 import com.xingren.excel.converter.read.IReadConverter;
+import com.xingren.excel.pojo.CellEntity;
 import com.xingren.excel.pojo.ExcelColumnAnnoEntity;
 
 /**
@@ -10,7 +11,7 @@ import com.xingren.excel.pojo.ExcelColumnAnnoEntity;
 public class DefaultReadConverter implements IReadConverter {
 
     @Override
-    public Object convert(ExcelColumnAnnoEntity entity, Class<?> clazz, Object cellValue,Object rowObj) {
+    public Object convert(ExcelColumnAnnoEntity entity, CellEntity cellEntity, Object rowObj) {
         throw new UnsupportedOperationException("该类属于注解默认参数,不可实例化");
     }
 }
