@@ -2,6 +2,7 @@ package com.xingren.excel.converter.read.impl;
 
 import com.xingren.excel.converter.read.IReadConverter;
 import com.xingren.excel.exception.ExcelConvertException;
+import com.xingren.excel.pojo.CellEntity;
 import com.xingren.excel.pojo.ExcelColumnAnnoEntity;
 
 import java.lang.reflect.InvocationTargetException;
@@ -16,15 +17,18 @@ import java.util.stream.Stream;
  * @author guang
  * @since 2020/2/11 2:44 下午
  */
-public class EnumReadConverter implements IReadConverter {
+public class EnumReadConverter implements IReadConverter<Object, Object> {
 
     @Override
-    public Object convert(ExcelColumnAnnoEntity entity, Class<?> clazz, Object cellValue, Object rowObj) {
+    public Object convert(ExcelColumnAnnoEntity entity, CellEntity cellEntity, Object rowObj) {
+        Class<?> clazz = rowObj.getClass();
         Class<?> fieldType = entity.getField().getType();
+        String cellValue = cellEntity.getCellValue();
         if (!fieldType.isEnum()) {
             throw new ExcelConvertException("类 " + clazz.getName() + " 中字段:"
                     + entity.getFiledName() + " 不是 Enum 类型!");
         }
+        // todo#guang 循环中会执行此方法,性能瓶颈,需要优化
         return getEnumConstant(fieldType).get(cellValue);
     }
 
@@ -41,11 +45,7 @@ public class EnumReadConverter implements IReadConverter {
                     try {
                         Object value = type.getMethod(method.getName()).invoke(obj);
                         map.put(value.toString(), obj);
-                    } catch (IllegalAccessException e) {
-                        e.printStackTrace();
-                    } catch (InvocationTargetException e) {
-                        e.printStackTrace();
-                    } catch (NoSuchMethodException e) {
+                    } catch (IllegalAccessException | InvocationTargetException | NoSuchMethodException e) {
                         e.printStackTrace();
                     }
                 }
