@@ -6,7 +6,6 @@ import com.xingren.excel.pojo.ExcelColumnAnnoEntity;
 import com.xingren.excel.util.DateUtil;
 import com.xingren.excel.util.ReflectorUtil;
 
-import java.lang.reflect.InvocationTargetException;
 import java.lang.reflect.Method;
 import java.time.OffsetDateTime;
 
@@ -17,21 +16,16 @@ import java.time.OffsetDateTime;
 public class OffSetDateTimeWriteConverter implements IWriteConverter {
 
     @Override
-    public Object convert(ExcelColumnAnnoEntity entity, Class<?> clazz, Object rowData) {
+    public Object convert(ExcelColumnAnnoEntity entity, Object rowData) {
+        Class<?> clazz = rowData.getClass();
         if (!OffsetDateTime.class.equals(entity.getField().getType())) {
             throw new ExcelConvertException("类 " + clazz.getName() + " 中字段:"
                     + entity.getFiledName() + " 不是 OffsetDateTime 类型!");
         }
 
         Method getMethod = ReflectorUtil.fromCache(clazz).getGetMethod(entity.getFiledName());
-        OffsetDateTime offsetDateTime = null;
-        try {
-            offsetDateTime = (OffsetDateTime) getMethod.invoke(rowData);
-        } catch (IllegalAccessException e) {
-            e.printStackTrace();
-        } catch (InvocationTargetException e) {
-            e.printStackTrace();
-        }
+
+        OffsetDateTime offsetDateTime = (OffsetDateTime) ReflectorUtil.invokeGetMethod(getMethod, rowData);
 
         return DateUtil.formatOffsetDateTime(offsetDateTime, entity.getDatePattern());
 
