@@ -1,30 +1,23 @@
 package com.xingren.excel.service.read;
 
-import com.xingren.excel.converter.read.impl.EnumReadConverter;
+import com.xingren.excel.converter.read.ReadConverterFactory;
+import com.xingren.excel.converter.read.impl.PrefixAndSuffixReadConverter;
 import com.xingren.excel.exception.ExcelException;
 import com.xingren.excel.pojo.CellEntity;
 import com.xingren.excel.pojo.ErrorInfoRow;
 import com.xingren.excel.pojo.ExcelColumnAnnoEntity;
 import com.xingren.excel.pojo.RowEntity;
 import com.xingren.excel.service.ExcelColumnService;
-import com.xingren.excel.util.DateUtil;
 import com.xingren.excel.util.ReflectorUtil;
-import org.apache.commons.lang3.ObjectUtils;
 import org.apache.commons.lang3.StringUtils;
 import org.apache.poi.ss.formula.functions.T;
-import org.apache.poi.ss.usermodel.Cell;
 
 import java.lang.reflect.Method;
-import java.time.LocalDateTime;
-import java.time.OffsetDateTime;
 import java.util.ArrayList;
 import java.util.List;
 import java.util.Map;
 import java.util.function.Function;
 import java.util.stream.Collectors;
-
-import static com.xingren.excel.ExcelConstant.DEFAULT_DATE_PATTREN;
-import static org.apache.poi.ss.usermodel.CellType.NUMERIC;
 
 /**
  * @author guang
@@ -35,6 +28,8 @@ public class ExcelReadService {
     private ReflectorUtil reflectorUtil;
 
     private Class<T> clazz;
+
+    private ReadConverterFactory readConverterFactory = new ReadConverterFactory();
 
     private ExcelReadService(Class clazz) {
         this.clazz = clazz;
@@ -62,123 +57,47 @@ public class ExcelReadService {
     private <T> void parseColumnEntityToRowObj(ExcelColumnAnnoEntity annoEntity, T rowObj, CellEntity cellEntity) {
         Map<String, Method> fieldNameSetMethodMap = reflectorUtil.getSetMethods();
 
-        // 如果在 Excel 中的字段在 POJO 中没有相应注解,跳过赋值
         if (null == annoEntity) {
             return;
         }
+
         Method setMethod = fieldNameSetMethodMap.get(annoEntity.getFiledName());
-        String columnValue = cellEntity.getCellValue();
         Class<?> fieldType = annoEntity.getField().getType();
-        if (null != annoEntity.getReadConverter()) {
-            Object fieldValue = annoEntity.getReadConverter().convert(annoEntity, clazz, columnValue, rowObj);
-            invokeSetMethod(rowObj, setMethod, fieldValue, fieldType, annoEntity, columnValue);
-            return;
-        }
-        // 前缀/后缀处理
+        String cellValue = null;
+        // 前缀/后缀处理(去除前缀后缀)
         if (StringUtils.isNotEmpty(annoEntity.getPrefix()) || StringUtils.isNotEmpty(annoEntity.getSuffix())) {
-            columnValue = StringUtils.substring(columnValue, annoEntity.getPrefix().length(),
-                    annoEntity.getPrefix().length() + columnValue.length() - annoEntity.getSuffix().length());
+            cellValue = new PrefixAndSuffixReadConverter().convert(annoEntity, cellEntity, rowObj);
+            // 去除前缀后缀后更新 cellValue 值
+            cellEntity.setCellValue(cellValue);
         }
-        if (String.class.equals(fieldType)) {
-            invokeSetMethod(rowObj, setMethod, columnValue, fieldType, annoEntity, columnValue);
-            return;
-        }
-        // Boolean 类型处理
-        if (Boolean.class.equals(fieldType)) {
-            processBoolean(rowObj, annoEntity, setMethod, columnValue, fieldType);
-            return;
-        }
-        // 元转分处理
-        if (annoEntity.getYuanToCent()) {
-            processYuanToCent(rowObj, setMethod, columnValue, fieldType, annoEntity);
-            return;
-        }
-        if (Integer.class.equals(fieldType) && !annoEntity.getYuanToCent()) {
-            invokeSetMethod(rowObj, setMethod, Integer.valueOf(columnValue), fieldType, annoEntity, columnValue);
-            return;
-        }
-        if (Long.class.equals(fieldType) && !annoEntity.getYuanToCent()) {
-            invokeSetMethod(rowObj, setMethod, Long.valueOf(columnValue), fieldType, annoEntity, columnValue);
-            return;
-        }
-        // 枚举处理
-        if (fieldType.isEnum()) {
-            // Enum 实例
-            Object enumObj = new EnumReadConverter().convert(annoEntity, clazz, columnValue, rowObj);
-            invokeSetMethod(rowObj, setMethod, enumObj, fieldType, annoEntity, columnValue);
-            return;
-        }
-        // OffsetDateTime 处理
-        if (OffsetDateTime.class.equals(fieldType)) {
-            if (StringUtils.isNotEmpty(columnValue)) {
-                Cell cell = cellEntity.getCell();
-                OffsetDateTime offsetDateTime = null;
-                if (cell.getCellTypeEnum().equals(NUMERIC)) {
-                    offsetDateTime = DateUtil.parseToOffsetDateTime(columnValue, DEFAULT_DATE_PATTREN);
-                } else {
-                    offsetDateTime = DateUtil.parseToOffsetDateTime(columnValue, annoEntity.getDatePattern());
-                }
-                invokeSetMethod(rowObj, setMethod, offsetDateTime, fieldType, annoEntity, columnValue);
-            }
-            return;
-        }
-        // LocalDateTime 处理
-        if (LocalDateTime.class.equals(fieldType)) {
-            Cell cell = cellEntity.getCell();
-            LocalDateTime localDateTime = null;
-            if (cell.getCellTypeEnum().equals(NUMERIC)) {
-                localDateTime = DateUtil.parseToLocalDateTime(columnValue, DEFAULT_DATE_PATTREN);
-            } else {
-                localDateTime = DateUtil.parseToLocalDateTime(columnValue, annoEntity.getDatePattern());
-            }
-            invokeSetMethod(rowObj, setMethod, localDateTime, fieldType, annoEntity, columnValue);
-            return;
-        }
+
+        // 以下基本数据类型的处理 ↓↓↓↓↓
+
+        Object value = readConverterFactory.getConverter(annoEntity).convert(annoEntity, cellEntity, rowObj);
+        invokeSetMethod(rowObj, setMethod, value, fieldType, annoEntity, cellValue);
     }
 
-    private <T> void processBoolean(T rowObj, ExcelColumnAnnoEntity annoEntity, Method setMethod, String columnValue,
-                                    Class<?> fieldType) {
-        String strToFalse = annoEntity.getStrToFalse();
-        String strToTrue = annoEntity.getStrToTrue();
-        if (StringUtils.isNotEmpty(strToFalse) && columnValue.equals(strToFalse) || "false".equals(columnValue)) {
-            invokeSetMethod(rowObj, setMethod, Boolean.FALSE, fieldType, annoEntity, columnValue);
-            return;
-        }
-        if (StringUtils.isNotEmpty(strToTrue) && columnValue.equals(strToTrue) || "true".equals(columnValue)) {
-            invokeSetMethod(rowObj, setMethod, Boolean.TRUE, fieldType, annoEntity, columnValue);
-            return;
-        }
-    }
-
-    private <T> void processYuanToCent(T rowObj, Method setMethod, String columnValue, Class<?> fieldType,
-                                       ExcelColumnAnnoEntity annoEntity) {
-        if (StringUtils.isBlank(columnValue)) {
-            return;
-        }
-        Double doubleValue = Double.valueOf(columnValue) * 100;
-        if (Integer.class.equals(fieldType)) {
-            Integer intValue = doubleValue.intValue();
-            invokeSetMethod(rowObj, setMethod, intValue, fieldType, annoEntity, columnValue);
-
-        }
-        if (Long.class.equals(fieldType)) {
-            Long longValue = doubleValue.longValue();
-            invokeSetMethod(rowObj, setMethod, longValue, fieldType, annoEntity, columnValue);
-        }
-    }
-
+    /**
+     * @param rowObj     当前 row 对象的 java 对象
+     * @param setMethod  当前 字段的 setXXX 方法
+     * @param setValue   当前字段值(已经有具体类型的值)
+     * @param setClass   当前字段类型
+     * @param annoEntity 当前字段上的注解
+     * @param cellValue  当前字段值(统一为 String 类型的形式)
+     * @return 范湖赋值后的 row 对象
+     */
     private <T> void invokeSetMethod(T rowObj, Method setMethod, Object setValue, Class<?> setClass,
-                                     ExcelColumnAnnoEntity annoEntity, String columnValue) {
+                                     ExcelColumnAnnoEntity annoEntity, String cellValue) {
         try {
             Class<?> fieldType = annoEntity.getField().getType();
-            // required 为 true 处理
+            // required 为 true 处理: 分为 枚举值 和 非枚举值 两种情况
             if (annoEntity.getRequired()) {
                 // 枚举值不合法
-                if (fieldType.isEnum() && StringUtils.isNotBlank(columnValue)) {
-                    throw new ExcelException("枚举值[" + columnValue + "]不合法");
+                if (fieldType.isEnum() && null == setValue) {
+                    throw new ExcelException("[" + annoEntity.getColumnName() + "]不合法");
                 }
                 // 空值
-                if (ObjectUtils.isEmpty(setValue) || "".equals(setValue)) {
+                if (null == setValue || "".equals(setValue)) {
                     throw new ExcelException("[" + annoEntity.getColumnName() + "]不能为空");
                 }
 
@@ -186,8 +105,8 @@ public class ExcelReadService {
                 // required 为 false 处理
                 // 如果是枚举类型
                 if (fieldType.isEnum()) {
-                    if (null == setValue && StringUtils.isNotBlank(columnValue)) {
-                        throw new ExcelException("枚举值[" + columnValue + "]不合法");
+                    if (null == setValue && StringUtils.isNotBlank(cellValue)) {
+                        throw new ExcelException("[" + cellValue + "]不合法");
                     }
                     if (null == setValue) {
                         return;
@@ -196,9 +115,10 @@ public class ExcelReadService {
             }
             ReflectorUtil.invokeSetMethod(rowObj, setMethod, setValue, setClass);
         } catch (Exception e) {
-            // 是否需要出错信息进行捕获
+            // 是否需要对出错信息进行捕获
             if (isErrorInfoRow(rowObj)) {
-                setErrotInfo((ErrorInfoRow) rowObj, e.getMessage());
+                setErrorInfo((ErrorInfoRow) rowObj, e.getMessage());
+                // 把异常信息栈打印出来,但是不主动抛出,因为继承了 ErrorInfoRow 类
                 e.printStackTrace();
             } else {
                 throw new ExcelException(e.getMessage());
@@ -206,10 +126,16 @@ public class ExcelReadService {
         }
     }
 
-    private <T> void setErrotInfo(ErrorInfoRow rowObj, String errorInfo) {
+    /**
+     * 设置出错信息,多个用换行符分割
+     */
+    private <T> void setErrorInfo(ErrorInfoRow rowObj, String errorInfo) {
         rowObj.setErrorInfo(errorInfo + "\n");
     }
 
+    /**
+     * 判断 该对象是否继承自 ErrorInfoRow
+     */
     private <T> boolean isErrorInfoRow(T rowObj) {
         return rowObj instanceof ErrorInfoRow;
     }
